refactor: separate auth and user routes
