+++ src/config/swagger.yaml
@@ -18,7 +18,8 @@ servers:
     description: production server
 
 paths:
-  # Users
+  # Authentication
+
   /activate-account:
     get:
       parameters:
@@ -29,7 +30,7 @@ paths:
             type: string
           description: Check your email after registration and enter JWT token which you got
       tags:
-        - User
+        - Authentication
       description: Activate user account after registration with JWT token
       responses:
         '200':
@@ -76,7 +77,7 @@ paths:
             type: string
           description: Enter your email address
       tags:
-        - User
+        - Authentication
       description: Verify email address and get email with JWT token
       responses:
         '200':
@@ -122,7 +123,7 @@ paths:
             type: string
           description: Enter user id
       tags:
-        - User
+        - Authentication
       description: Activate new user email
       responses:
         '200':
@@ -163,58 +164,10 @@ paths:
         '500':
           description: Internal Server Error.
 
-  /user-details:
-    get:
-      security:
-        - bearerAuth: []
-      parameters:
-        - in: query
-          name: accessToken
-          required: true
-          schema:
-            type: string
-          description: Enter access token which you got after authorization
-      tags:
-        - User
-      description: Fetch user details
-      responses:
-        '200':
-          description: User details fetched successfully
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  _id:
-                    type: string
-                  name:
-                    type: string
-                  email:
-                    type: string
-                example:
-                  _id: 62e9ec0df44f549d9b153423
-                  name: elon musk
-                  email: spacex@gmail.com
-        '401':
-          description: Provided JWT token is invalid.
-        '404':
-          description: User account not found
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  message:
-                    type: string
-                example:
-                  message: User not found
-        '500':
-          description: Internal Server Error.
-
   /register-user:
     post:
       tags:
-        - User
+        - Authentication
       description: Register new user
       requestBody:
         required: true
@@ -275,7 +228,7 @@ paths:
   /google-auth:
     post:
       tags:
-        - User
+        - Authentication
       description: Authenticate with google account and get JWT token
       requestBody:
         required: true
@@ -311,7 +264,7 @@ paths:
   /authorization:
     post:
       tags:
-        - User
+        - Authentication
       description: Log in ang get JWT token
       requestBody:
         required: true
@@ -374,6 +327,55 @@ paths:
                   message: User not found
         '422':
           description: Unprocessed Entity
+        '500':
+          description: Internal Server Error.
+
+  # Users
+  /user-details:
+    get:
+      security:
+        - bearerAuth: []
+      parameters:
+        - in: query
+          name: accessToken
+          required: true
+          schema:
+            type: string
+          description: Enter access token which you got after authorization
+      tags:
+        - User
+      description: Fetch user details
+      responses:
+        '200':
+          description: User details fetched successfully
+          content:
+            application/json:
+              schema:
+                type: object
+                properties:
+                  _id:
+                    type: string
+                  name:
+                    type: string
+                  email:
+                    type: string
+                example:
+                  _id: 62e9ec0df44f549d9b153423
+                  name: elon musk
+                  email: spacex@gmail.com
+        '401':
+          description: Provided JWT token is invalid.
+        '404':
+          description: User account not found
+          content:
+            application/json:
+              schema:
+                type: object
+                properties:
+                  message:
+                    type: string
+                example:
+                  message: User not found
         '500':
           description: Internal Server Error.
