+++ src/controllers/index.ts
@@ -1,1 +1,2 @@
 export * from './user-controller'
+export * from './auth-controller'
