+++ src/routes/index.ts
@@ -1,1 +1,2 @@
 export { default as userRouter } from './user'
+export { default as authRouter } from './auth'
