+++ src/routes/user.ts
@@ -1,48 +1,17 @@
+import { changeUserSchema, passwordSchema, idSchema } from 'schemas'
 import { validateRequestSchema } from 'middlewares'
 import express, { RequestHandler } from 'express'
 import { uploadUserImage } from 'utils'
 import {
-  userAccountActivation,
   changeUserCredentials,
-  activateNewUserEmail,
-  verifyUserEmail,
   changePassword,
   getUserDetails,
   uploadUserImg,
-  authorization,
-  registerUser,
-  googleAuth,
 } from 'controllers'
-import {
-  googleUserSchema,
-  changeUserSchema,
-  passwordSchema,
-  userSchema,
-  authSchema,
-  idSchema,
-} from 'schemas'
 
 const router = express.Router()
 
-router.post('/google-auth', googleUserSchema, validateRequestSchema, googleAuth)
-
-router.post('/authorization', authSchema, validateRequestSchema, authorization)
-
-router.get('/activate-account', userAccountActivation)
-
-router.get('/verify-email', verifyUserEmail)
-
 router.get('/user-details', getUserDetails)
-
-router.get('/activate-new-user-email', activateNewUserEmail)
-
-router.post(
-  '/register-user',
-  passwordSchema,
-  userSchema,
-  validateRequestSchema,
-  registerUser
-)
 
 router.post(
   '/change-password',
