+++ src/server.ts
@@ -1,8 +1,8 @@
 import { swaggerMiddleware, authMiddleware } from 'middlewares'
 import express, { RequestHandler } from 'express'
+import { userRouter, authRouter } from 'routes'
 import SwaggerUI from 'swagger-ui-express'
 import { connectToMongo } from 'config'
-import { userRouter } from 'routes'
 import dotenv from 'dotenv'
 import cors from 'cors'
 
@@ -19,6 +19,7 @@ server.use(express.static('public'))
 
 server.use(authMiddleware as RequestHandler)
 
+server.use(authRouter)
 server.use(userRouter)
 
 server.listen(process.env.SERVER_PORT, () => {
